bob-generative-api/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
bcrypt==4.1.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
ecdsa==0.19.0
fastapi==0.88.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==1.36.1
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.2
PyJWT==2.3.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.22.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.18.2
yarl==1.9.4
